# public/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Farmart API",
    "description": "REST API contract for Farmart — Thai online grocery store.\n\n## Color Rules (mirrored from design system)\n- **Amber** (`#F4B223`) browse/navigation actions\n- **Green-600** (`#16a34a`) purchase actions: add-to-cart, checkout, membership\n\n## Auth\nAll cart and checkout endpoints require a session cookie (`farmart-session`). Register or login first to obtain one.\n\n## Price\nAll `price` fields are in **Thai Baht (THB)**, stored as integers (สตางค์-free). Always render via `formatPrice()` on the client.",
    "version": "0.1.0",
    "contact": {
      "name": "Farmart Dev",
      "email": "dev@farmart.th"
    }
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local dev" },
    { "url": "https://farmart.vercel.app", "description": "Production" }
  ],
...
  "paths": {
...
    "/api/search": {
      "get": {
        "tags": ["search"],
        "summary": "Search products",
        "description": "Full-text search across product name and category. Delegates to `/api/products?q=` internally — provided as a convenience alias.",
        "operationId": "searchProducts",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "description": "Search term (min 2 characters)",
            "schema": { "type": "string", "minLength": 2 }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "default": 10, "maximum": 50 }
          }
        ],
        "responses": {
          "200": {
            "description": "Matching products",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProductListResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "farmart-session",
        "description": "HttpOnly session cookie issued by `/api/auth/login` or `/api/auth/register`"
      }
    },
    "schemas": {
...
      "Product": {
        "type": "object",
        "required": ["id", "name", "category", "price", "discount", "image", "rating", "reviewCount", "isNew"],
        "properties": {
          "id":          { "type": "string",  "example": "1" },
          "name":        { "type": "string",  "example": "บร็อคโคลี่ออร์แกนิค" },
          "category":    { "type": "string",  "example": "ผัก" },
          "price":       { "type": "integer", "description": "Price in THB", "example": 99 },
          "discount":    { "type": "integer", "description": "Discount percentage 0–100", "minimum": 0, "maximum": 100, "example": 20 },
          "image":       { "type": "string",  "format": "uri", "example": "https://images.unsplash.com/photo-example" },
          "rating":      { "type": "number",  "format": "float", "minimum": 0, "maximum": 5, "example": 4.5 },
          "reviewCount": { "type": "integer", "minimum": 0, "example": 128 },
          "isNew":       { "type": "boolean", "example": false }
        }
      },
...
      "ProductListResponse": {
        "type": "object",
        "required": ["data", "meta"],
        "properties": {
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Product" }
          },
          "meta": {
            "type": "object",
            "required": ["total", "page", "limit", "totalPages"],
            "properties": {
              "total":      { "type": "integer", "example": 20 },
              "page":       { "type": "integer", "example": 1 },
              "limit":      { "type": "integer", "example": 20 },
              "totalPages": { "type": "integer", "example": 1 }
            }
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": ["code", "message"],
        "properties": {
          "code":    { "type": "string", "example": "NOT_FOUND" },
          "message": { "type": "string", "example": "ไม่พบสินค้าที่ร้องขอ" }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid request parameters or body",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "code": "VALIDATION_ERROR", "message": "ข้อมูลไม่ถูกต้อง" }
          }
        }
      },
...
      "InternalError": {
        "description": "Unexpected server error",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "code": "INTERNAL_ERROR", "message": "เกิดข้อผิดพลาด กรุณาลองใหม่อีกครั้ง" }
          }
        }
      }
    }
  }
}